# openapi.yaml
openapi: 3.0.3
info:
  title: EasyBooking API
  version: 1.0.0
  description: REST API for the EasyBooking endterm project.
servers:
  - url: http://127.0.0.1:3000
paths:
...
  /api/hotels:
    get:
      summary: List hotels
      parameters:
        - in: query
          name: page
          schema: { type: integer, minimum: 1 }
        - in: query
          name: limit
          schema: { type: integer, minimum: 1 }
        - in: query
          name: q
          schema: { type: string }
        - in: query
          name: city
          schema: { type: string }
        - in: query
          name: sort
          schema: { type: string, example: rating_desc }
      responses:
        '200':
          description: Hotels list
    post:
      summary: Create hotel (admin)
      security: [{ cookieAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HotelCreateInput'
      responses:
        '201':
          description: Created
        '403':
          description: Forbidden
  /api/hotels/{id}:
    get:
      summary: Get hotel by ID
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Hotel
        '404':
          description: Not found
    put:
      summary: Update hotel (admin)
      security: [{ cookieAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HotelUpdateInput'
      responses:
        '200':
          description: Updated
    delete:
      summary: Delete hotel (admin)
      security: [{ cookieAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Deleted
  /api/hotels/{id}/rate:
    post:
      summary: Rate hotel (auth)
      security: [{ cookieAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [score]
              properties:
                score:
                  type: integer
                  minimum: 1
                  maximum: 5
      responses:
        '200':
          description: Rating saved
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: easybook.sid
  parameters:
    IdPath:
      in: path
      name: id
      required: true
      schema:
        type: string
      description: MongoDB ObjectId
...
  schemas:
...
    HotelCreateInput:
      type: object
      required:
        - title
        - description
        - location
        - address
        - price_per_night
        - rating
        - available_rooms
        - amenities
      properties:
        title: { type: string }
        description: { type: string }
        location: { type: string }
        address: { type: string }
        price_per_night: { type: number }
        rating: { type: number }
        available_rooms: { type: integer }
        amenities:
          type: array
          items: { type: string }
        imageUrl: { type: string }
    HotelUpdateInput:
      type: object
      properties:
        title: { type: string }
        description: { type: string }
        location: { type: string }
        address: { type: string }
        price_per_night: { type: number }
        rating: { type: number }
        available_rooms: { type: integer }
        amenities:
          type: array
          items: { type: string }
        imageUrl: { type: string }
...